+++ internal/processing/ory/access_rule_processor_test.go
@@ -274,87 +274,6 @@ var _ = Describe("Access Rule Processor", func() {
 			})
 		})
 
-		When("rule exists and and rule path is different", func() {
-			It("should create new rule and delete old rule", func() {
-				// given
-				noop := []*gatewayv1beta1.Authenticator{
-					{
-						Handler: &gatewayv1beta1.Handler{
-							Name: "noop",
-						},
-					},
-				}
-
-				noopRule := GetRuleFor("newPath", ApiMethods, []*gatewayv1beta1.Mutator{}, noop)
-				rules := []gatewayv1beta1.Rule{noopRule}
-
-				apiRule := GetAPIRuleFor(rules)
-
-				rule := rulev1alpha1.Rule{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							processing.OwnerLabelv1alpha1: fmt.Sprintf("%s.%s", apiRule.ObjectMeta.Name, apiRule.ObjectMeta.Namespace),
-						},
-					},
-					Spec: rulev1alpha1.RuleSpec{
-						Match: &rulev1alpha1.Match{
-							URL: fmt.Sprintf("<http|https>://%s<%s>", ServiceHost, "oldPath"),
-						},
-					},
-				}
-
-				vs := networkingv1beta1.VirtualService{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							processing.OwnerLabelv1alpha1: fmt.Sprintf("%s.%s", apiRule.ObjectMeta.Name, apiRule.ObjectMeta.Namespace),
-						},
-					},
-				}
-
-				scheme := runtime.NewScheme()
-				err := rulev1alpha1.AddToScheme(scheme)
-				Expect(err).NotTo(HaveOccurred())
-				err = networkingv1beta1.AddToScheme(scheme)
-				Expect(err).NotTo(HaveOccurred())
-				err = gatewayv1beta1.AddToScheme(scheme)
-				Expect(err).NotTo(HaveOccurred())
-
-				client := fake.NewClientBuilder().WithScheme(scheme).WithObjects(&rule, &vs).Build()
-				processor := ory.NewAccessRuleProcessor(GetTestConfig())
-
-				// when
-				result, err := processor.EvaluateReconciliation(context.TODO(), client, apiRule)
-
-				// then
-				Expect(err).To(BeNil())
-				Expect(result).To(HaveLen(2))
-
-				createResultMatcher := PointTo(MatchFields(IgnoreExtras, Fields{
-					"Action": Equal("create"),
-					"Obj": PointTo(MatchFields(IgnoreExtras, Fields{
-						"Spec": MatchFields(IgnoreExtras, Fields{
-							"Match": PointTo(MatchFields(IgnoreExtras, Fields{
-								"URL": Equal("<http|https>://myService.myDomain.com<newPath>"),
-							})),
-						}),
-					})),
-				}))
-
-				deleteResultMatcher := PointTo(MatchFields(IgnoreExtras, Fields{
-					"Action": Equal("delete"),
-					"Obj": PointTo(MatchFields(IgnoreExtras, Fields{
-						"Spec": MatchFields(IgnoreExtras, Fields{
-							"Match": PointTo(MatchFields(IgnoreExtras, Fields{
-								"URL": Equal("<http|https>://myService.myDomain.com<oldPath>"),
-							})),
-						}),
-					})),
-				}))
-
-				Expect(result).To(ContainElements(createResultMatcher, deleteResultMatcher))
-			})
-		})
-
 	})
 
 	When("multiple handler", func() {
