+++ internal/processing/virtual_service_processor.go
@@ -9,11 +9,11 @@ import (
 
 type VirtualServiceProcessor struct {
 	Creator           VirtualServiceCreator
-	OathkeeperSvc     string
-	OathkeeperSvcPort uint32
-	CorsConfig        *CorsConfig
-	AdditionalLabels  map[string]string
-	DefaultDomainName string
+	oathkeeperSvc     string
+	oathkeeperSvcPort uint32
+	corsConfig        *CorsConfig
+	additionalLabels  map[string]string
+	defaultDomainName string
 }
 
 type VirtualServiceCreator interface {
